+++ src/types/fieldArray.ts
@@ -14,6 +14,7 @@ export type UseFieldArrayProps<
   name: TFieldArrayName;
   keyName?: TKeyName;
   control?: Control<TFieldValues>;
+  shouldUnregister?: boolean;
 };
 
 export type FieldArrayWithId<