+++ src/types/form.ts
@@ -84,6 +84,7 @@ export type UseFormProps<
   resolver: Resolver<TFieldValues, TContext>;
   context: TContext;
   shouldFocusError: boolean;
+  shouldUnregister: boolean;
   criteriaMode: CriteriaMode;
 }>;
 
@@ -144,6 +145,12 @@ export type UseFormRegister<TFieldValues extends FieldValues> = <
   name: TFieldName,
   options?: RegisterOptions<TFieldValues, TFieldName>,
 ) => UseFormRegisterReturn;
+
+export type UseFormSetFocus<TFieldValues extends FieldValues> = <
+  TFieldName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>
+>(
+  name: TFieldName,
+) => void;
 
 export type UseFormGetValues<TFieldValues extends FieldValues> = {
   (): UnpackNestedValue<TFieldValues>;
@@ -395,6 +402,19 @@ export type UseFormUnregister<TFieldValues extends FieldValues> = (
   > & { keepValue?: boolean; keepDefaultValue?: boolean; keepError?: boolean },
 ) => void;
 
+export type UseFormInternalUnregister<TFieldValues extends FieldValues> = (
+  name?: FieldPath<TFieldValues> | FieldPath<TFieldValues>[],
+  options?: Omit<
+    KeepStateOptions,
+    | 'keepIsSubmitted'
+    | 'keepSubmitCount'
+    | 'keepValues'
+    | 'keepDefaultValues'
+    | 'keepErrors'
+  > & { keepValue?: boolean; keepDefaultValue?: boolean; keepError?: boolean },
+  notify?: boolean,
+) => void;
+
 export type UseFormHandleSubmit<TFieldValues extends FieldValues> = <
   TSubmitFieldValues extends FieldValues = TFieldValues
 >(
@@ -415,15 +435,16 @@ export type WatchInternal<TFieldValues> = (
   | FieldPathValue<FieldValues, InternalFieldName>
   | FieldPathValues<FieldValues, InternalFieldName[]>;
 
-export type GetFormIsDirty = <TName extends InternalFieldName, TData>(
+export type GetIsDirty = <TName extends InternalFieldName, TData>(
   name?: TName,
   data?: TData,
 ) => boolean;
 
 export type Control<TFieldValues extends FieldValues = FieldValues> = {
+  shouldUnmountUnregister?: boolean;
   isWatchAllRef: React.MutableRefObject<boolean>;
   watchFieldsRef: React.MutableRefObject<InternalNameSet>;
-  getFormIsDirty: GetFormIsDirty;
+  getIsDirty: GetIsDirty;
   fieldArrayDefaultValuesRef: FieldArrayDefaultValues;
   formStateRef: React.MutableRefObject<FormState<TFieldValues>>;
   formStateSubjectRef: React.MutableRefObject<
@@ -459,6 +480,7 @@ export type Control<TFieldValues extends FieldValues = FieldValues> = {
   defaultValuesRef: React.MutableRefObject<DefaultValues<TFieldValues>>;
   watchInternal: WatchInternal<TFieldValues>;
   register: UseFormRegister<TFieldValues>;
+  unregister: UseFormUnregister<TFieldValues>;
 };
 
 export type WatchObserver<TFieldValues> = (
@@ -483,6 +505,7 @@ export type UseFormReturn<TFieldValues extends FieldValues = FieldValues> = {
   unregister: UseFormUnregister<TFieldValues>;
   control: Control<TFieldValues>;
   register: UseFormRegister<TFieldValues>;
+  setFocus: UseFormSetFocus<TFieldValues>;
 };
 
 export type UseFormStateProps<TFieldValues> = Partial<{