+++ src/useFieldArray.ts
@@ -39,6 +39,7 @@ export const useFieldArray = <
   control,
   name,
   keyName = 'id' as TKeyName,
+  shouldUnregister,
 }: UseFieldArrayProps<
   TFieldValues,
   TFieldArrayName,
@@ -49,7 +50,7 @@ export const useFieldArray = <
   const {
     isWatchAllRef,
     watchFieldsRef,
-    getFormIsDirty,
+    getIsDirty,
     watchSubjectRef,
     fieldArraySubjectRef,
     fieldArrayNamesRef,
@@ -61,6 +62,8 @@ export const useFieldArray = <
     validFieldsRef,
     fieldsWithValidationRef,
     fieldArrayDefaultValuesRef,
+    unregister,
+    shouldUnmountUnregister,
   } = control || methods.control;
 
   const [fields, setFields] = React.useState<
@@ -89,7 +92,11 @@ export const useFieldArray = <
     );
 
   const getCurrentFieldsValues = () => {
-    const values = get(getFieldsValues(fieldsRef, defaultValuesRef), name, []);
+    const values = get(
+      getFieldsValues(fieldsRef, defaultValuesRef.current),
+      name,
+      [],
+    );
 
     return mapIds<TFieldValues, TKeyName>(
       get(fieldArrayDefaultValuesRef.current, name, []).map(
@@ -223,7 +230,7 @@ export const useFieldArray = <
     }
 
     formStateSubjectRef.current.next({
-      isDirty: getFormIsDirty(name, omitKey(updatedFieldArrayValues)),
+      isDirty: getIsDirty(name, omitKey(updatedFieldArrayValues)),
       errors: formStateRef.current.errors as FieldErrors<TFieldValues>,
       isValid: formStateRef.current.isValid,
     });
@@ -398,7 +405,11 @@ export const useFieldArray = <
 
     watchSubjectRef.current.next({
       name,
-      value: get(getFieldsValues(fieldsRef, defaultValuesRef), name, []),
+      value: get(
+        getFieldsValues(fieldsRef, defaultValuesRef.current),
+        name,
+        [],
+      ),
     });
 
     focusNameRef.current &&
@@ -435,8 +446,8 @@ export const useFieldArray = <
     !get(fieldsRef.current, name) && set(fieldsRef.current, name, []);
 
     return () => {
-      fieldArrayDefaultValuesRef.current = getFieldsValues(fieldsRef);
       fieldArraySubscription.unsubscribe();
+      (shouldUnmountUnregister || shouldUnregister) && unregister(name);
     };
   }, []);